refactor(settings): move SettingRecord + definitions to base module (ADR-020)
Completes the ADR-020 rollout: SettingRecord entity, its QueryDefinition
and ExportDefinition move to Granit.Settings. EF Core configuration and
store stay in the .EntityFrameworkCore project.
