+++ src/Granit.Settings.EntityFrameworkCore/Extensions/SettingsEntityFrameworkCoreHostApplicationBuilderExtensions.cs
@@ -1,10 +1,5 @@
-using Granit.DataExchange.Extensions;
-using Granit.QueryEngine.Extensions;
 using Granit.Settings.Definitions;
-using Granit.Settings.EntityFrameworkCore.Entities;
-using Granit.Settings.EntityFrameworkCore.Exports;
 using Granit.Settings.EntityFrameworkCore.Internal;
-using Granit.Settings.EntityFrameworkCore.Queries;
 using Granit.Settings.Values;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
@@ -50,10 +45,6 @@ public static class SettingsEntityFrameworkCoreHostApplicationBuilderExtensions
             sp.GetRequiredService<EfCoreSettingStore<TDbContext>>()));
         builder.Services.Replace(ServiceDescriptor.Singleton<ISettingStoreWriter>(sp =>
             sp.GetRequiredService<EfCoreSettingStore<TDbContext>>()));
-
-        // Query + Export definitions (ADR-020: owned by the base module).
-        builder.Services.AddQueryDefinition<SettingRecord, SettingRecordQueryDefinition>();
-        builder.Services.AddExportDefinition<SettingRecord, SettingRecordExportDefinition>();
 
         return builder;
     }
