+++ src/Granit.Settings.EntityFrameworkCore/Internal/SettingRecordConfiguration.cs
@@ -1,4 +1,4 @@
-using Granit.Settings.EntityFrameworkCore.Entities;
+using Granit.Settings.Domain;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
