+++ tests/Granit.Settings.EntityFrameworkCore.Tests/EfCoreSettingStoreTests.cs
@@ -6,7 +6,7 @@
 // =============================================================================
 
 using Granit.Settings.Definitions;
-using Granit.Settings.EntityFrameworkCore.Entities;
+using Granit.Settings.Domain;
 using Granit.Settings.EntityFrameworkCore.Extensions;
 using Granit.Settings.EntityFrameworkCore.Internal;
 using Granit.Settings.Values;
