+++ tests/Granit.Settings.EntityFrameworkCore.Tests/SettingRecordConfigurationTests.cs
@@ -4,7 +4,7 @@
 // Verifies table mapping, column constraints, and unique index definition.
 // =============================================================================
 
-using Granit.Settings.EntityFrameworkCore.Entities;
+using Granit.Settings.Domain;
 using Granit.Settings.EntityFrameworkCore.Extensions;
 using Granit.Settings.EntityFrameworkCore.Internal;
 using Microsoft.EntityFrameworkCore;
