+++ tests/Granit.Settings.EntityFrameworkCore.Tests/SettingsEfCoreDiRegistrationTests.cs
@@ -1,4 +1,4 @@
-using Granit.Settings.EntityFrameworkCore.Entities;
+using Granit.Settings.Domain;
 using Granit.Settings.EntityFrameworkCore.Extensions;
 using Granit.Settings.EntityFrameworkCore.Internal;
 using Granit.Settings.Values;
